feat(client device auth plugin): commit initial code for the auth plugin
Initial commit for the auth plugin.
This is based on the POC.
The change adds an erlang plugin based on the template provided by the emqx team.
It has adds the driver code and native code to be called form the erlang code.
The integration is demonstrated by end to end plumbing the on_client_authenticate emqx hook from erlang all the way to C++ code.