Create an CDA auth plugin for EMQX (#4)
* feat(client device auth plugin): commit initial code for the auth plugin

Initial commit for the auth plugin.
This is based on the POC.
The change adds an erlang plugin based on the template provided by the emqx team.
It has adds the driver code and native code to be called form the erlang code.
The integration is demonstrated by end to end plumbing the on_client_authenticate emqx hook from erlang all the way to C++ code.

* feat(client device auth plugin): plumb emqx hook callbacks to c++ code

This change builds on top of the skeleton code for the auth plugin. It plumbs the emqx auth hooks all the way to the c++ code. All the client login to connect to CDA via IPC can live in the cpp code.
The tests are missing right now and will follow. The change has been tested manually.

The PEM information for the client cert is received in the on_connect hook and is remembered in the erlang process's dictionary. This is later retrieved and used along with client id to call into the cpp code. These together will constitute the mqtt credentials. username and password are skipped for now and can be added later.

Co-authored-by: Michael Dombrowski <mdombro@amazon.com>