refactor: copy loaded_plugins in copy_config script (#89)
Copying of loaded_plugins has been moved to copy_config.cmd script. Eventually copy_config.cmd will be replaced, but in the meantime, this commit has a fix for our component not starting up with a clean work directory, due to copying loaded_plugins to a nested dir that wouldn't exist.